refactor: pull out types and helper functions
